Extract shared parallelFor utilities
